+++ backend/src/enums/lent.type.enum.ts
