+++ package.json
@@ -4,5 +4,9 @@
   "main": "index.js",
   "repository": "https://github.com/innovationacademy-kr/42cabi.git",
   "author": "--replace-all <gyuwonlee.olivia@gmail.com>",
-  "license": "MIT"
+  "license": "MIT",
+  "dependencies": {
+    "@nestjs/typeorm": "^9.0.1",
+    "typeorm": "^0.3.9"
+  }
 }
